Add query

+++ graphql/types/strategy_type.rb
@@ -3,9 +3,9 @@ module Types
   StrategyType = GraphQL::ObjectType.define do
     name 'Strategy'
     field :action, types.String
-    field :entry, types.Float
-    field :target, types.Float
-    field :stoploss, types.Float
+    field :entry, types.Int
+    field :target, types.Int
+    field :stoploss, types.Int
     field :note, types.String
   end
 end
